+++ platform/macos/src/MGLMapViewDelegate.h
@@ -244,6 +244,18 @@ NS_ASSUME_NONNULL_BEGIN
 #pragma mark Selecting Annotations
 
 /**
+ Returns a Boolean value indicating whether the shape annotation can be selected.
+ 
+ If the return value is `YES`, the user can select the annotation by clicking
+ on it. If the delegate does not implement this method, the default value is `YES`.
+ 
+ @param mapView The map view that has selected the annotation.
+ @param annotation The object representing the shape annotation.
+ @return A Boolean value indicating whether the annotation can be selected.
+ */
+- (BOOL)mapView:(MGLMapView *)mapView shapeAnnotationIsEnabled:(MGLShape *)annotation;
+
+/**
  Tells the delegate that one of its annotations has been selected.
 
  You can use this method to track changes to the selection state of annotations.